Move private dandi.files classes to _private.py
